2.1.2 - 修复能量雨工具/主账户登录失败 部分模块独立及优化 截图权限申请方法优化
