+++ README.md
@@ -115,6 +115,13 @@
 
 [comment]: <> (Version history only shows last 3 versions)
 
+# v2.1.2
+###### 2021/05/19
+* `修复` 能量雨收集工具高概率出现截图权限申请失败的问题 _[`issue #467`](https://github.com/SuperMonster003/Ant-Forest/issues/467)_
+* `修复` 支付宝应用控件变更导致主账户登录失败的问题
+* `优化` 独立无障碍服务模块(a11y)及本地存储扩展模块(storage)
+* `优化` 截图权限申请扩展方法在多线程等极端条件下的适应性
+
 # v2.1.1
 ###### 2021/05/17
 * `新增` 能量雨收集工具 (简易) (/tools/energy-rain-harvesting.js)
@@ -134,13 +141,6 @@
 * `优化` 项目代码中所有尾逗号的处理 (Add when multiline)
 * `优化` 清除好友森林页面统计能量数据时对于用户昵称的依赖 _[`issue #228`](https://github.com/SuperMonster003/Ant-Forest/issues/228)_  _[`#223`](https://github.com/SuperMonster003/Ant-Forest/issues/223)_  _[`#218`](https://github.com/SuperMonster003/Ant-Forest/issues/218)_
 * `优化` 增加支付宝手势锁监听器并在触发后适时结束脚本 _[`issue #452`](https://github.com/SuperMonster003/Ant-Forest/issues/452)_
-
-# v2.0.5
-###### 2021/03/16
-* `修复` 解锁功能配置参数无法获取默认值的问题 _[`issue #444`](https://github.com/SuperMonster003/Ant-Forest/issues/444)_
-* `修复` 排行榜可点击样本位于屏幕底部区域时可能出现的点击偏移问题 _[`issue #443`](https://github.com/SuperMonster003/Ant-Forest/issues/443)_
-* `优化` 运行提示对话框放弃任务时增加不再提示选项 _[`issue #445`](https://github.com/SuperMonster003/Ant-Forest/issues/445)_
-* `优化` 清除配置工具参数文件中混入的非必要参数数据
 
 ##### 更多版本历史可参阅
 * [CHANGELOG-2.md](https://github.com/SuperMonster003/Ant-Forest/blob/master/documents/CHANGELOG-2.md)  ( 2.x 版本 )
