+++ documents/CHANGELOG-2.md
@@ -1,6 +1,13 @@
 ******
 ### 版本历史 - 2.x
 ******
+# v2.1.2
+###### 2021/05/19
+* `修复` 能量雨收集工具高概率出现截图权限申请失败的问题 _[`issue #467`](https://github.com/SuperMonster003/Ant-Forest/issues/467)_
+* `修复` 支付宝应用控件变更导致主账户登录失败的问题
+* `优化` 独立无障碍服务模块(a11y)及本地存储扩展模块(storage)
+* `优化` 截图权限申请扩展方法在多线程等极端条件下的适应性
+
 # v2.1.1
 ###### 2021/05/17
 * `新增` 能量雨收集工具 (简易) (/tools/energy-rain-harvesting.js)
