+++ tools/energy-balls-recognition.js
@@ -2,6 +2,7 @@ require('../modules/ext-images').load();
 require('../modules/ext-global').load();
 require('../modules/ext-device').load();
 require('../modules/ext-dialogs').load();
+require('../modules/ext-storages').load();
 
 dialogsx.builds([
     '能量球识别测试工具',
@@ -25,9 +26,8 @@ function _() {
     let $_cfg = {};
     let $_sto = {};
     let _wins = [];
-    let _mod_sto = require('../modules/mod-storage');
     Object.assign($_sto, {
-        cfg: _mod_sto.create('af_cfg'),
+        cfg: storagesx.create('af_cfg'),
         def: require('../modules/mod-default-config'),
     });
